Be more careful about path handling for Snakefiles, job wrappers, etc.
There was a bug in calling `basename()` on HTCondor executable job wrappers
because using something like `workflow/wrapper.sh` would get flattened
to `wrapper.sh` and added as the `executable` argument for HTCondor's
CEDAR transfer mechanism. Because the path doesn't exist, HTCondor
can't transfer the file.

This commit fixes the bug, but also pulls some of the other stuff into
separate helper functions so that they're easier to mock/test.